Added CASGatewayCheckCookie
Looks for the presence of cookie defined by CASGatewayCheckCookie.  If it isn't found, anonymous access to the resource is granted.  If the cookie is present, it indicates that this browser has accessed the CAS server and therefore has the potential to be in an authenticated state.  Thus the request is redirected following the CAS gateway protocol.